% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D15_vectors}
\alias{D15_vectors}
\title{D15 vectors}
\format{
A data frame with 16 rows and 4 variables:
\describe{
  \item{u}{}
  \item{v}{}
  \item{su}{}
  \item{sv}{}
  ...
}
}
\usage{
D15_vectors
}
\description{
The vectors in color space of the 15 caps that make up the Panel D15 color vision test.
}
\keyword{datasets}
