% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norm_values}
\alias{norm_values}
\title{Norm values [dB]}
\format{

}
\usage{
norm_values
}
\description{
The norm values for photoreceptor-specific temporal contrast sensitivity [dB] for a 40-year old subject. These were re-calculated from the data published in TVST: Huchzermeyer et. al. Photoreceptor-Specific Loss of Perifoveal Temporal Contrast Sensitivity in Retinitis Pigmentosa. Transl Vis Sci Technol 2020:9(6):27.
}
\keyword{datasets}
