% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vingrys.R
\name{vingrys}
\alias{vingrys}
\title{Calculates scores for the Panel D15 according to Vingrys et al.}
\usage{
vingrys(caps)
}
\arguments{
\item{caps}{a vector of cap numbers}
}
\value{
a list with parameters
}
\description{
Calculates scores for the Panel D15 according to Vingrys et al.
}
